+++ static/app/views/settings/project/dynamicSampling/rule.spec.tsx
@@ -4,7 +4,6 @@ import {ServerSideSamplingStore} from 'sentry/stores/serverSideSamplingStore';
 import {SamplingRuleOperator} from 'sentry/types/sampling';
 
 import {Rule} from './rule';
-import {uniformRule} from './testUtils.spec';
 
 export const samplingBreakdownTitle = 'Transaction Breakdown';
 
@@ -18,7 +17,7 @@ describe('Server-Side Sampling - Rule', function () {
       <Rule
         operator={SamplingRuleOperator.IF}
         hideGrabButton={false}
-        rule={uniformRule}
+        rule={TestStubs.DynamicSamplingConfig().uniformRule}
         onEditRule={() => {}}
         onDeleteRule={() => {}}
         onActivate={() => {}}
@@ -50,7 +49,7 @@ describe('Server-Side Sampling - Rule', function () {
       <Rule
         operator={SamplingRuleOperator.IF}
         hideGrabButton={false}
-        rule={{...uniformRule, active: true}}
+        rule={{...TestStubs.DynamicSamplingConfig().uniformRule, active: true}}
         onEditRule={() => {}}
         onDeleteRule={() => {}}
         onActivate={() => {}}
